Cleaning up the routes
- Removed the initialisation endpoint, mainly due to the fact that the compiler started screaming at me and since it was meant to be a quick fix rather than a longer term architecture thing I have decided to remove it. It seemed like it was bad practice anyways
