+++ orders_api/src/graphql/mod.rs
@@ -1,0 +1,5 @@
+
+mod query;
+mod mutation;
+pub use query::QueryRoot;
+pub use mutation::MutationRoot;
