+++ orders_api/src/lib.rs
@@ -1,1 +1,3 @@
 pub mod models;
+pub mod routes;
+pub mod graphql;
