+++ orders_api/src/models/mod.rs
@@ -4,6 +4,7 @@ mod error;
 mod fry;
 mod meal;
 mod menu;
+mod schema;
 mod size;
 
 pub use burger::Burger;
@@ -13,9 +14,5 @@ pub use error::CustomError;
 pub use fry::Fries;
 pub use meal::Meal;
 pub use menu::Menu;
+pub use schema::ShopSchema;
 pub use size::Size;
-
-mod query;
-mod mutation;
-pub use query::QueryRoot;
-pub use mutation::MutationRoot;
