+++ orders_api/src/models/schema.rs
@@ -1,0 +1,4 @@
+use async_graphql::{EmptySubscription, Schema};
+use crate::graphql::{MutationRoot, QueryRoot};
+
+pub type ShopSchema = Schema<QueryRoot, MutationRoot, EmptySubscription>;
