+++ orders_api/src/routes/mod.rs
@@ -1,0 +1,4 @@
+
+pub mod graphql;
+
+pub use graphql::*;
